ci(workspace): test

+++ apps/web/scripts/deploy.sh
@@ -44,8 +44,8 @@ deploy() {
         remote_plan "$CONFIG"
     fi
     rm "$TF_DIR/$PLAN"
-    cleanup
     upload_assets_to_cdn
+    cleanup
     generate_cdn_dns_entry
     generate_domain_mapping_dns_entry
 }
@@ -109,7 +109,7 @@ image() {
             exit 1
         fi
         NEXT_PUBLIC_PROJECT_TYPE="$CONFIG" nx build "$APP" --skip-nx-cache
-        INPUT_GITHUB_TOKEN="$INPUT_GITHUB_TOKEN" INPUT_IMAGES="$INPUT_IMAGES" INPUT_TAGS="sha-$SHA" nx docker "$APP" --skip-nx-cache
+        INPUT_GITHUB_TOKEN="$INPUT_GITHUB_TOKEN" INPUT_IMAGES="$INPUT_IMAGES" INPUT_TAGS="sha-$SHA" nx docker "$APP" --configuration=ci --skip-nx-cache
     fi
     docker push "$TAGGED_IMAGE"
 }
@@ -126,27 +126,46 @@ cleanup() {
     REPO_NAME=$($TF output repository_id | tr -d '"')
     REPO="$ARTIFACT_REGION-$REGISTRY/$PROJECT/$REPO_NAME"
     IMAGE="$ARTIFACT_REGION-$REGISTRY/$PROJECT/$REPO_NAME/$REPO_NAME"
-    IMAGES=$(gcloud artifacts docker images list "$REPO" --include-tags --sort-by=CREATE_TIME | tail -n +2)
-    echo "$IMAGES"
-    IMAGE_COUNT=$(echo "$IMAGES" | wc -l | tr -d ' ')
-    i="$IMAGE_COUNT_THRESHOLD"
+    TIMER_THRESHOLD=60
+    START=$(date +%s)
     set +e
-    while [ $i -lt "$IMAGE_COUNT" ]; do
-        DIGEST=$(echo "$IMAGES" | sed -n "${i}p" | awk '{print $2}')
-        TAG=$(echo "$IMAGES" | sed -n "${i}p" | awk '{print $3}')
-        case $TAG in
-        sha-*)
-            purple "gcloud artifacts docker images delete $IMAGE:$TAG --delete-tags"
-            RESULT="$(echo "" | gcloud artifacts docker images delete "$IMAGE:$TAG" --delete-tags)"
-            echo "$RESULT"
-            ;;
-        *)
-            purple "gcloud artifacts docker images delete $IMAGE@$DIGEST --delete-tags"
-            RESULT="$(echo "" | gcloud artifacts docker images delete "$IMAGE@$DIGEST" --delete-tags)"
-            echo "$RESULT"
-            ;;
-        esac
-        i=$((i + 1))
+    while true; do
+        NOW=$(date +%s)
+        ELAPSED=$((NOW - START))
+        if [ $ELAPSED -ge $TIMER_THRESHOLD ]; then
+            purple "timer threshold exceeded, probably stuck in an infinite loop, breaking out"
+            break
+        fi
+        IMAGES=$(gcloud artifacts docker images list "$REPO" --include-tags --sort-by=CREATE_TIME | tail -n +2)
+        echo "$IMAGES"
+        IMAGE_COUNT=$(echo "$IMAGES" | wc -l | tr -d ' ')
+        i="$IMAGE_COUNT_THRESHOLD"
+        while [ $i -lt "$IMAGE_COUNT" ]; do
+            DIGEST=$(echo "$IMAGES" | sed -n "${i}p" | awk '{print $2}')
+            TAG=$(echo "$IMAGES" | sed -n "${i}p" | awk '{print $3}')
+            case $TAG in
+            sha-*)
+                purple "gcloud artifacts docker images delete $IMAGE:$TAG --delete-tags"
+                RESULT="$(echo "" | gcloud artifacts docker images delete "$IMAGE:$TAG" --delete-tags)"
+                echo "$RESULT"
+                ;;
+            *)
+                purple "gcloud artifacts docker images delete $IMAGE@$DIGEST --delete-tags"
+                RESULT="$(echo "" | gcloud artifacts docker images delete "$IMAGE@$DIGEST" --delete-tags)"
+                echo "$RESULT"
+                ;;
+            esac
+            i=$((i + 1))
+        done
+        IMAGES=$(gcloud artifacts docker images list "$REPO" --include-tags --sort-by=CREATE_TIME | tail -n +2)
+        echo "$IMAGES"
+        IMAGE_COUNT=$(echo "$IMAGES" | wc -l | tr -d ' ')
+        if [ "$IMAGE_COUNT" -gt "$IMAGE_COUNT_THRESHOLD" ]; then
+            purple "More images than threshold, running loop again"
+        else
+            purple "Number of images now matches threshold"
+            break
+        fi
     done
     set -e
 }
